[af-207] test: NotProd 수정으로 인한 test 코드 수정

+++ src/test/java/com/drunkenlion/alcoholfriday/domain/admin/item/application/AdminItemServiceTest.java
@@ -89,7 +89,7 @@ public class AdminItemServiceTest {
     private final String productName = "1000억 막걸리 프리바이오";
     private final BigDecimal productPrice = BigDecimal.valueOf(10000);
     private final Long productQuantity = 1000L;
-    private final Long alcohol = 10L;
+    private final Double alcohol = 10D;
     private final String ingredient = "쌀(국내산), 밀(국내산), 누룩, 정제수";
     private final Long sweet = 10L;
     private final Long sour = 10L;
@@ -103,7 +103,7 @@ public class AdminItemServiceTest {
     private final String modifyProductName = "1000억 막걸리 프리바이오 수정";
     private final BigDecimal modifyProductPrice = BigDecimal.valueOf(1000);
     private final Long modifyProductQuantity = 100L;
-    private final Long modifyAlcohol = 1L;
+    private final Double modifyAlcohol = 1D;
     private final String modifyIngredient = "쌀(국내산), 밀(국내산), 누룩, 정제수 수정";
     private final Long modifySweet = 1L;
     private final Long modifySour = 1L;
