+++ src/test/java/com/drunkenlion/alcoholfriday/domain/cart/application/CartServiceTest.java
@@ -61,7 +61,7 @@ class CartServiceTest {
     private final BigDecimal price = new BigDecimal(50000);
     private final String info = "이 상품은 테스트 상품입니다.";
     private final Long quantity = 10L;
-    private final Long alcohol = 17L;
+    private final Double alcohol = 17D;
     private final String ingredient = "알콜, 누룩 등등...";
     private final Long sweet = 10L;
     private final Long sour = 10L;
@@ -80,7 +80,7 @@ class CartServiceTest {
     private final BigDecimal price2 = new BigDecimal(100_000);
     private final String info2 = "이 상품은 테스트 상품2 입니다.";
     private final Long quantity2 = 10L;
-    private final Long alcohol2 = 17L;
+    private final Double alcohol2 = 17D;
     private final String ingredient2 = "알콜, 누룩 등등...";
     private final Long sweet2 = 10L;
     private final Long sour2 = 10L;
